requirements.txt
# Web framework
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1

# Forms and validation
WTForms==3.0.1

# Proxmox integration
proxmoxer==2.0.1

# SSH and file transfer
paramiko==3.3.1

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Security
Werkzeug==2.3.7

# Additional dependencies
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
bcrypt==4.0.1
cryptography==41.0.7
